requirements.txt
django==4.2
djangorestframework==3.14.0
celery==5.3.1
redis==4.3.4
flower==1.2.0
psycopg2-binary==2.9.2
